refactor: move getCardAttribute function to a separate module for improved organization; update imports across the engine to reflect the new structure
